Use componentDidMount instead of componentWillMount
componentDidMount is only called when the component actually mounts. This means it will only be called when rendering on the client since the server will never actually mount any components anywhere when rendering.

componentWillMount, on the other hand, will be called before the component mounts when rendering on the client and also be called on the server where window will not be defined.

componentDidMount is the more appropriate place for the code which uses the DOM and sets event handlers on window.